templates/starter/agent/requirements.txt
strands-agents[mcp]
bedrock-agentcore
uv
